Fixed warnings, improved boundstree debug

+++ application/shapeLibrary.cpp
@@ -2,7 +2,7 @@
 #include "../engine/math/mathUtil.h"
 
 namespace Library {
-	double g = (1.0 + sqrt(5.0)) / 2.0;
+	float g = (1.0 + sqrt(5.0)) / 2.0;
 
 	Vec3f icosahedronVertices[]{
 		Vec3f(0, 0.5, g / 2), Vec3f(0, 0.5, -g / 2), Vec3f(0, -0.5, -g / 2), Vec3f(0, -0.5, g / 2),
